refactor code into hex design pattern
