+++ go.mod
@@ -3,7 +3,6 @@ module github.com/mike_jacks/pizza_co
 go 1.22
 
 require (
-	github.com/bxcodec/faker/v4 v4.0.0-beta.3
 	github.com/joho/godotenv v1.5.1
 	google.golang.org/grpc v1.65.0
 	google.golang.org/protobuf v1.34.2
@@ -13,5 +12,5 @@ require (
 	golang.org/x/net v0.28.0 // indirect
 	golang.org/x/sys v0.24.0 // indirect
 	golang.org/x/text v0.17.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20240808171019-573a1156607a // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20240812133136-8ffd90a71988 // indirect
 )
