+++ order_management_service/proto/v1/order_management_service.proto
@@ -1,0 +1,11 @@
+syntax = "proto3";
+
+package order_management_service.proto.v1;
+
+import "order_management_service/proto/v1/order_management.proto";
+
+option go_package = "github.com/mike_jacks/pizza_co/order_management_service/ports/grpc/v1;order_management_v1_pb";
+
+service OrderManagementService {
+    rpc PlaceOrder(OrderRequest) returns (stream OrderResponse) {}
+}
