--- protos/inventory/v1/inventory.proto
@@ -1,39 +1,0 @@
-syntax = "proto3";
-
-package github.com.mike_jacks.pizza_co.protos.inventory.v1;
-
-import "order_management/v1/order_management.proto";
-
-option go_package = "github.com/mike_jacks/pizza_co/generated/inventory/v1;inventory_v1_pb";
-
-message InventoryCheckRequest {
-    repeated order_management.v1.Pizza pizzas = 1;
-}
-
-message InventoryCheckResponse {
-    string message = 1;
-    uint32 error_code = 2;
-    bool is_available = 3;
-}
-
-message UpdateInventoryRequest {
-    UpdateType update_type = 1;
-    repeated order_management.v1.Pizza.Topping toppings = 2;
-    repeated PizzaCrust pizza_crust = 3;
-
-    message PizzaCrust {
-        order_management.v1.Pizza.CrustType crust_type = 1;
-        order_management.v1.Pizza.Size crust_size = 2;
-    }
-
-    enum UpdateType {
-        UPDATE_TYPE_UNKNOWN = 0;
-        ADD = 1;
-        DELETE = 2;
-    }
-}
-
-message UpdateInventoryResponse {
-    string message = 1;
-    uint32 error_code = 2;
-}
