--- protos/inventory/v1/inventory_service.proto
@@ -1,12 +1,0 @@
-syntax = "proto3";
-
-package github.com.mike_jacks.pizza_co.protos.inventory.v1;
-
-import "inventory/v1/inventory.proto";
-
-option go_package = "github.com/mike_jacks/pizza_co/generated/inventory/v1;inventory_v1_pb";
-
-service InventoryService {
-    rpc CheckInventory(InventoryCheckRequest) returns (InventoryCheckResponse) {}
-    rpc UpdateInventory(UpdateInventoryRequest) returns (UpdateInventoryResponse) {}
-}
