--- protos/order_management/v1/order_management_service.proto
@@ -1,11 +1,0 @@
-syntax = "proto3";
-
-package github.com.mike_jacks.pizza_co.protos.order_management.v1;
-
-import "order_management/v1/order_management.proto";
-
-option go_package = "github.com/mike_jacks/pizza_co/generated/order_management/v1;order_management_v1_pb";
-
-service OrderManagementService {
-    rpc PlaceOrder(OrderRequest) returns (stream OrderResponse) {}
-}
