--- protos/payment/v1/payment.proto
@@ -1,5 +1,0 @@
-syntax = "proto3";
-
-package github.com.mike_jacks.pizza_co.protos.payment.v1;
-
-option go_package = "github.com/mike_jacks/pizza_co/generated/payment/v1;payment_v1_pb";
